pyproject.toml
[project]
name = "chatpdfv3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "chromadb>=0.5.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
]
embeddings = [
    "sentence-transformers>=3.0.1",
]

[project.scripts]
chatpdf = "chatpdfv3.interfaces.cli:main"

[tool.setuptools.packages.find]
include = ["chatpdfv3*"]
exclude = ["files*"]
